README page headings & text content
