+++ app/views/stash_engine/resources/prepare_readme.html.erb
@@ -11,7 +11,7 @@
 <% end %>
 
 <!-- map(&:attributes) one way for translating ActiveRecord field type to json at React side adequately -->
-<%= react_component('components/ReadMe', {
+<%= react_component('containers/ReadMe', {
   readmeUrl: @file_location
 }) %>
